PG dla Ulamka
wektor do dokonczenia

+++ Gauss/Macierze.java
@@ -40,7 +40,7 @@ public class Macierze {
                 r = losujR();
                 macierzF[i][j] = (float) r/65536;
                 macierzD[i][j] = (double) r/65536;
-                macierzU[i][j] = new Ulamek(r/65536);
+                macierzU[i][j] = new Ulamek((double) r/65536);
             }
         }
     }
